adding simple transliteration to dictionary
