+++ data/dict/prefixes/entries/Hb.json
@@ -179,5 +179,7 @@
     "בָ",
     "בָּ֛"
   ],
-  "frequency": 15847
+  "frequency": 15847,
+  "transliteration_en": "BE",
+  "transliteration_es": "BE"
 }
