+++ data/dict/prefixes/entries/Hc.json
@@ -98,5 +98,7 @@
     "וָֽ",
     "וָ֫"
   ],
-  "frequency": 51272
+  "frequency": 51272,
+  "transliteration_en": "VE",
+  "transliteration_es": "VE"
 }
