+++ data/dict/prefixes/entries/Hd.json
@@ -55,5 +55,7 @@
     "הַ",
     "הֶֽ"
   ],
-  "frequency": 24060
+  "frequency": 24060,
+  "transliteration_en": "HA",
+  "transliteration_es": "HA"
 }
