+++ data/dict/prefixes/entries/Hk.json
@@ -72,5 +72,7 @@
     "כְ֭",
     "כְ֨"
   ],
-  "frequency": 2973
+  "frequency": 2973,
+  "transliteration_en": "KE",
+  "transliteration_es": "KE"
 }
