+++ data/dict/prefixes/entries/Hl.json
@@ -201,5 +201,7 @@
     "לַֽ",
     "לִ֬"
   ],
-  "frequency": 20852
+  "frequency": 20852,
+  "transliteration_en": "LE",
+  "transliteration_es": "LE"
 }
