+++ data/dict/prefixes/entries/Hm.json
@@ -57,5 +57,7 @@
     "מִ֠",
     "מִ֝"
   ],
-  "frequency": 6414
+  "frequency": 6414,
+  "transliteration_en": "MI",
+  "transliteration_es": "MI"
 }
