+++ data/dict/prefixes/entries/Ht.json
@@ -20,5 +20,7 @@
     "את",
     "אֵת"
   ],
-  "frequency": 0
+  "frequency": 0,
+  "transliteration_en": "Et",
+  "transliteration_es": "Et"
 }
