+++ tools/bani/__init__.py
@@ -1,0 +1,9 @@
+"""
+Bani - Hebrew Transliteration System
+Provides simple transliteration for Hebrew text to English and Spanish.
+"""
+
+from .transliterate import BaniTransliterator
+
+__version__ = "2.0.0"
+__all__ = ["BaniTransliterator"]
